requirements.txt
flask==3.0.0
langchain==0.3.0
langchain-openai==0.2.0
werkzeug==3.0.1
python-dotenv==1.0.0

